Refine cluster example for tutorial
