+++ examples/tutorial_config.json
@@ -2,6 +2,6 @@
   "input_zarr": "sample_data.zarr",
   "output_zarr": "sample_data.zarr",
   "input_group": "raw",
-  "output_group": "smoothed_subprocess_config",
+  "output_group": "smoothed_subprocess",
   "sigma": 5
 }
